RE-1169 Specify container base image
In most cases, it's irrelevant which base image we use for creating a
container, however an ops-fabric test requires an ubuntu:17.10
container and subsequent tests call python3.6, which is not available
in the standard ubuntu base container image we use for all tests.

This review introduces the parameter `SLAVE_CONTAINER_BASE_IMAGE`,
allowing projects to specify which Ubuntu-flavored container base
image is used. Additionally, we adjust the Dockerfile to allow the
base image to overridden.

NOTE: There are two issues with this implementation at present.
Overriding FROM in the Dockerfile requires --build-arg to be passed
to docker build, which is not exposed via Jenkins' docker.build(). As
such, we have to shell out and invoke docker build ourselves.
Additionally, after doing testing on this PR, it seems that the
docker pipeline plugin does not like the changes to the Dockerfile
(see [1] for more info). [1] has been resolved and released, however
we are using a custom build of this plugin (1.11).

[1] https://issues.jenkins-ci.org/browse/JENKINS-46105